Continue integer type refactoring: use std::size_t for sizes of (and indices into) containers, std::int64_t for storing in Variant, and plain int for everything else

+++ OpenSolid/Core/LazyCollection.definitions.hpp
@@ -66,7 +66,7 @@ namespace opensolid
         bool
         isEmptyDefaultImpl() const;
 
-        std::int64_t
+        std::size_t
         sizeDefaultImpl() const;
     public:
         const TDerived&
@@ -81,7 +81,7 @@ namespace opensolid
         bool
         isEmpty() const;
 
-        std::int64_t
+        std::size_t
         size() const;
 
         template <class TPredicate>
