+++ OpenSolid/Core/LazyCollection/FilteredCollection.definitions.hpp
@@ -69,7 +69,7 @@ namespace opensolid
             bool
             isEmpty() const;
 
-            std::int64_t
+            std::size_t
             size() const;
         };
